Position Diagram (temporary)
This is roughly what i was trying to do with the position values...All numbers without a letter next to it have the position value (number,'x') and the starting point gets a value (0,"x") when your start and (0,"y") when you end. Branching is handled by rules that transform, say, (1,b) ---> (5,a). I'm not the biggest fan of this strategy, but it'll do for the moment....